requirements.txt
apache-beam==2.1.0
astral==1.4
avro==1.8.2
backports.weakref==1.0.post1
beam==0.4.0
bleach==2.0.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
crcmod==1.7
cycler==0.10.0
dill==0.2.7.1
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.90.4
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
geopy==1.11.0
google-apitools==0.5.14
google-auth==1.1.0
google-auth-httplib2==0.0.2
google-cloud==0.27.0
google-cloud-bigquery==0.27.0
google-cloud-bigtable==0.27.0
google-cloud-core==0.27.1
google-cloud-datastore==1.3.0
google-cloud-dns==0.27.0
google-cloud-error-reporting==0.27.0
google-cloud-language==0.29.0
google-cloud-logging==1.3.0
google-cloud-monitoring==0.27.0
google-cloud-pubsub==0.28.3
google-cloud-resource-manager==0.27.0
google-cloud-runtimeconfig==0.27.0
google-cloud-spanner==0.27.0
google-cloud-speech==0.29.0
google-cloud-storage==1.4.0
google-cloud-translate==1.2.0
google-cloud-videointelligence==0.27.2
google-cloud-vision==0.27.0
google-gax==0.15.15
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
googledatastore==7.0.1
googlemaps==2.5.1
grpc-google-iam-v1==0.11.3
grpcio==1.6.0
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
Markdown==2.6.9
matplotlib==2.0.2
mgrs==1.3.5
mock==2.0.0
monotonic==1.3
numpy==1.13.1
oauth2client==4.1.2
pandas==0.20.3
pbr==3.1.1
ply==3.10
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.4.0
psutil==5.3.1
pyasn1==0.3.5
pyasn1-modules==0.1.4
pygeocoder==1.2.5
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
scipy==0.19.1
seaborn==0.8.1
six==1.11.0
subprocess32==3.2.7
tenacity==4.4.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.6
tensorflow-transform==0.1.10
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.12.2
